Moves test form to appropriate folder
